% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{augustus}
\alias{augustus}
\title{wrapper to run AUGUSTUS}
\usage{
augustus(species, input, output = NULL, args = "", ...)
}
\arguments{
\item{species}{string. an identifier for the species model to use, such as \code{'chicken'}}

\item{input}{string. path to the input file, which contains the gene sequence(s) in fasta format.}

\item{output}{string. path to the output file. if \code{NULL} output is writen into standard output}

\item{args}{character. additional parameters passed on to AUGUSTUS}

\item{...}{other parameters passed on to \code{\link[=system2]{system2()}}, from \code{stdout} to the end}
}
\value{
result of \code{\link[=system2]{system2()}} or \code{NULL} if \code{input} doesn't exist
}
\description{
wrapper to run AUGUSTUS
}
\examples{
\dontrun{
    augustus('human', 'software/augustus/examples/example.fa');
    augustus('human', 'software/augustus/examples/hsackI10.fa.gz')

    augustus('human', 'software/augustus/examples/example.fa', 'test.gff');
}


}
