% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-fasta.R
\name{augustus_gff_to_fasta}
\alias{augustus_gff_to_fasta}
\title{convert gff file produced by augustus to fasta file storing protein}
\usage{
augustus_gff_to_fasta(input, output)
}
\arguments{
\item{output}{string. path to output file}

\item{input.}{string. path to input file}
}
\value{
\code{NULL}
}
\description{
convert gff file produced by augustus to fasta file storing protein
}
