% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{compare_sra_paired}
\alias{compare_sra_paired}
\title{compare fastqc result of original and trimmed sequence for paired ends}
\usage{
compare_sra_paired(id)
}
\arguments{
\item{id}{string. sra accession, such as \code{'SRR6148275'}}
}
\value{
\code{NULL}
}
\description{
compare fastqc result of original and trimmed sequence for paired ends
}
\examples{
\dontrun{
    compare_sra_paired('SRR6148275')
}

}
\keyword{internal}
