% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{compare_sra_single}
\alias{compare_sra_single}
\title{compare fastqc result of original and trimmed sequence for single ends}
\usage{
compare_sra_single(id)
}
\arguments{
\item{id}{string. sra accession, such as \code{'SRR029421'}}
}
\value{
\code{NULL}
}
\description{
compare fastqc result of original and trimmed sequence for single ends
}
\examples{
\dontrun{
    compare_sra_single('SRR029421')
}

}
\keyword{internal}
