% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paup.R
\name{fasta_to_nexus}
\alias{fasta_to_nexus}
\title{convert fasta to nexus data block.}
\usage{
fasta_to_nexus(fasta, datatype, missing = "?", gap = "-")
}
\arguments{
\item{fasta}{tibble. see \strong{Value} of \code{\link[biozhuoer:read_fasta]{biozhuoer::read_fasta()}}. Sequence
must be aligned (in the same length). \code{'U'} would be replaced by
\code{'X'} since paup doesn't recognized it}

\item{datatype}{string. what kind of data is you fasta. (\code{'dna'}, \code{'rna'},
\code{'protein'}, or \code{'standard'})}

\item{missing}{string. symbol used for missing data}

\item{gap}{string. symbol used for gap}
}
\value{
character. A Nexus data block. Note that begining '#nexus' isn't
included.
}
\description{
internal function for \code{\link[=write_paup]{write_paup()}}
}
\details{
the return value will be printed when call at console, if the file
is big, it will take a while, and you can't do anything. Consider
\code{fasta_to_nexus() -> x} to avoid automatically printing.

Alternatively, you can use seaview But it may take several seconds if each
sequence lies on only one line in the original file, because seaview always
force the line width less than 80.
}
\examples{
fasta <- biozhuoer::read_fasta(system.file('extdata', 'EOG090F05Z3.fasta', package = 'mcapomorphy'))
fasta_to_nexus(fasta, 'protein');

}
\keyword{internal}
