% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{fastqc_path}
\alias{fastqc_path}
\title{return the full absolute path of the fastqc result}
\usage{
fastqc_path(name)
}
\arguments{
\item{name}{string. name of the sequence file, such as \code{'SRR6148275_2'}}
}
\value{
string.
}
\description{
return the full absolute path of the fastqc result
}
\examples{
fastqc_path('SRR6148275_2')

}
\keyword{internal}
