% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_assembly}
\alias{fetch_assembly}
\title{download NCBI assembly webpage for a vector of given assessions}
\usage{
fetch_assembly(assessions)
}
\arguments{
\item{assessions}{character.}
}
\value{
\code{NULL}
}
\description{
download NCBI assembly webpage for a vector of given assessions
}
