% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_genome}
\alias{fetch_genome}
\title{download NCBI genome webpage for a vector of given id}
\usage{
fetch_genome(id)
}
\arguments{
\item{assessions}{character.}
}
\value{
\code{NULL}
}
\description{
download NCBI genome webpage for a vector of given id
}
