% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_TSA}
\alias{fetch_TSA}
\title{download NCBI WGS TSA assembly webpage for a vector of given ids}
\usage{
fetch_TSA(ids)
}
\arguments{
\item{ids}{character.}
}
\value{
\code{NULL}
}
\description{
download NCBI WGS TSA assembly webpage for a vector of given ids
}
