% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-fasta.R
\name{hamstr_from_augustus}
\alias{hamstr_from_augustus}
\title{prepare hamstr input file from augustus gff}
\usage{
hamstr_from_augustus(input, output)
}
\arguments{
\item{output}{string. path to output file}

\item{input.}{string. path to input file}
}
\value{
\code{NULL}
}
\description{
prepare hamstr input file from augustus gff
}
\details{
convert gff file produced by augustus to fasta file storing protein
}
