% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-fasta.R
\name{hamstr_from_oases}
\alias{hamstr_from_oases}
\title{prepare hamstr input file from oases transcript}
\usage{
hamstr_from_oases(input, output)
}
\arguments{
\item{output}{string. path to output file}

\item{input.}{string. path to input file}
}
\value{
\code{NULL}
}
\description{
prepare hamstr input file from oases transcript
}
