% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-fasta.R
\name{hamstr_out_to_fasta}
\alias{hamstr_out_to_fasta}
\title{convert hamstr output file to FASTA format}
\usage{
hamstr_out_to_fasta(input, output)
}
\arguments{
\item{output}{string. path to output file}

\item{input.}{character. path to input file, can be more than one.
non-existent ones are ignored.}
}
\value{
\code{NULL}
}
\description{
convert hamstr output file to FASTA format
}
\details{
if a OG exists multiple times, only the first one would be reserved
}
