% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_assembly}
\alias{parse_assembly}
\title{parse assembly metadata in xml format}
\usage{
parse_assembly(path)
}
\arguments{
\item{path}{\code{x} of \code{\link[xml2:read_xml]{xml2::read_xml()}}. assembly metadata in XML format}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
\enumerate{
\item species
\item id
\item similarity logical scalar.
\item partial logical scalar.
\item gca_dir
\item gcf_dir
}
}
\description{
parse assembly metadata in xml format
}
\section{to do}{
 unittest
}

\examples{
NULL

}
