% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_avibase}
\alias{parse_avibase}
\title{parse wikipedia species webpage}
\usage{
parse_avibase(path)
}
\arguments{
\item{path}{string. path to the HTML file.}
}
\value{
data_frame
\enumerate{
\item order
\item family
\item genus
\item species
}
}
\description{
parse wikipedia species webpage
}
\examples{
parse_avibase('data-raw/Avibase-1FDDABDB0D4421F9.html')

}
