% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_genome}
\alias{parse_genome}
\title{parse NCBI genome webpage}
\usage{
parse_genome(id)
}
\arguments{
\item{id}{string.}
}
\value{
data_frame
\enumerate{
\item species
\item number
\item dna
\item rna
\item protein
}
}
\description{
parse NCBI genome webpage
}
