% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{replace_ext}
\alias{replace_ext}
\title{replace extension name}
\usage{
replace_ext(filename, new_ext = "")
}
\description{
extension name can be any number of \code{\\w}
}
\section{to do}{
 unittest
}

\examples{
replace_ext('a.txt', '.md')

}
\keyword{internal}
