% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{view_sra_paired}
\alias{view_sra_paired}
\title{view fastqc result of original SRA sequence for paired ends}
\usage{
view_sra_paired(id)
}
\arguments{
\item{id}{string. sra accession, such as \code{'SRR6148275'}}
}
\value{
\code{NULL}
}
\description{
view fastqc result of original SRA sequence for paired ends
}
\examples{
\dontrun{
    view_sra_paired('SRR6148275')
}

}
\keyword{internal}
