% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{view_sra_single}
\alias{view_sra_single}
\title{view fastqc result of original SRA sequence for single ends}
\usage{
view_sra_single(id)
}
\arguments{
\item{id}{string. sra accession, such as \code{'SRR029421'}}
}
\value{
\code{NULL}
}
\description{
view fastqc result of original SRA sequence for single ends
}
\examples{
\dontrun{
    view_sra_single('SRR029421')
}

}
\keyword{internal}
