% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc-RNA-seq.R
\name{file_url}
\alias{file_url}
\title{browse fastqc html output view RStduio Server}
\usage{
file_url(path)
}
\arguments{
\item{path}{string. full absolute path of the file}
}
\value{
string.
}
\description{
browse fastqc html output view RStduio Server

return the url of a file on the workstation
}
\section{package options embedded in aves}{

options(mcapomorphy.wd = '/path/to/wd')
options(mcapomorphy.url = 'http://hostname:port')
}

\examples{
file_url(fastqc_path('SRR6148275_2'))

}
\keyword{internal}
