% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_avibase}
\alias{parse_avibase}
\title{parse avibase species webpage}
\usage{
parse_avibase(path)
}
\arguments{
\item{path}{string. path to the HTML file.}
}
\value{
data_frame
\enumerate{
\item order
\item family
\item genus
\item species
\item id
}
}
\description{
parse avibase species webpage
}
\details{
you have to download a Chinese webpage: note \code{lang=ZH} in \code{https://avibase.bsc-eoc.org/species.jsp?lang=ZH&amp;avibaseid=}
}
\examples{
parse_avibase('data-raw/01761045A529DC4A')


}
