% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_avibase}
\alias{parse_avibase}
\title{parse wikipedia species webpage}
\usage{
parse_avibase(id)
}
\arguments{
\item{id}{string. Avibase id}
}
\value{
data_frame
\enumerate{
\item order
\item family
\item genus
\item species
}
}
\description{
parse wikipedia species webpage
}
\examples{
aves::parse_avibase('900C590B5EEE5157')

}
