% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_genome}
\alias{parse_genome}
\title{parse NCBI genome webpage}
\usage{
parse_genome(path)
}
\arguments{
\item{path}{string. path to the HTML file.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
\enumerate{
\item species
\item number
\item dna
\item rna
\item protein
}
}
\description{
parse NCBI genome webpage
}
\examples{
parse_genome(system.file('extdata', 'genome55342.html', package = 'mcapomorphy'))
parse_genome(system.file('extdata', 'genome2793.html', package = 'mcapomorphy'))

}
